move interdeps fixtures
